requirements.txt
# FastAPI e dependências web
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# CORS (já vem integrado no FastAPI, não precisa instalar separado)

# Validação e tipos
email-validator==2.1.0

# Utilitários
python-dateutil==2.8.2

# Desenvolvimento e testes
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Cache (opcional, para fase 9)
redis==5.0.1
hiredis==2.3.2

# WebSocket (para fase 8)
websockets==12.0